Move CSP logic into app/csp.py.
